rollback unnecessary-change and add some doc
