+++ src/streamlink/plugins/douyu.py
@@ -19,12 +19,28 @@ from urllib.parse import urlparse
 
 from streamlink.plugin import Plugin, pluginmatcher
 from streamlink.plugin.api import validate
-from streamlink.session.http import TLSPartialVerifyAdapter
+from streamlink.session.http import TLSSecLevel1Adapter
 from streamlink.stream.http import HTTPStream
 from streamlink.utils.parse import parse_qsd
 
 
 log = logging.getLogger(__name__)
+
+
+class TLSPartialVerifyAdapter(TLSSecLevel1Adapter):
+    """
+    The base class SSLContextAdapter resets ssl_context.check_hostname to True when sending a request, so overriding get_ssl_context does not take effect.
+
+    Instead, assert_hostname is used to influence urllib3's internal decision logic,
+    ultimately achieving the effect of check_hostname=False to skip certificate hostname verification,
+    while still retaining certificate chain-related validation."
+    """
+
+    def init_poolmanager(self, *args, **kwargs):
+        kwargs["ssl_context"] = self.get_ssl_context()
+        kwargs["assert_hostname"] = False
+        return super().init_poolmanager(*args, **kwargs)
+
 
 @pluginmatcher(
     re.compile(
